requirements.txt
python-dotenv~=1.0.1
uvicorn~=0.23.2
fastapi~=0.110.2
langserve~=0.2.1

langchain~=0.2.5
pydantic==1.10.15